Adds black haunting wolf
Fixes PortraitRepository, adds NPCSystem and
necessary adjustments. I reworked Character related
structure.
Closes #9

+++ core/src/main/java/io/github/mufca/libgdx/datastructure/character/jsondata/PrimaryStatisticsData.java
@@ -5,12 +5,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record PrimaryStatisticsData(
-    @JsonProperty("strength") double strength,
-    @JsonProperty("dexterity") double dexterity,
-    @JsonProperty("constitution") double constitution,
-    @JsonProperty("intelligence") double intelligence,
-    @JsonProperty("wisdom") double wisdom,
-    @JsonProperty("charisma") double charisma
+    @JsonProperty("strength") float strength,
+    @JsonProperty("dexterity") float dexterity,
+    @JsonProperty("constitution") float constitution,
+    @JsonProperty("intelligence") float intelligence,
+    @JsonProperty("wisdom") float wisdom,
+    @JsonProperty("charisma") float charisma
 ) {
 
 }
